+++ src/anonymizers/types.ts
@@ -1,0 +1,3 @@
+import {BaseAnonymizer} from './base/base-anonymizer';
+
+export type Anonymizer = BaseAnonymizer;
