+++ src/cli/commands/anon-col/helpers/process-column.ts
@@ -1,5 +1,5 @@
 import ora from 'ora';
-import {Anonymizer} from '../../../../anonymizers';
+import {Anonymizer} from '../../../../anonymizers/types';
 import {Processor} from '../../../../processors/base-processor/base-processor';
 import {getProcessor} from './get-processor';
 import {isUserConfirmed} from './is-user-confirmed';
